+++ CHANGELOG.md
@@ -7,6 +7,7 @@ To install or update LODA, please follow the [installation instructions](https:/
 * Fix range computation for `dgr`
 * Fix IE bug occuring when using step count > 1 and `max` on loop counter
 * Fix argument check for `seq` for programs with offset
+* Fix minimization/optimization loop
 
 ## v25.6.27